requirements.txt
torch
transformers
streamlit
opencv-python
pillow
numpy
